update: 응답 DTO에서 평가 점수 정수부만 반환

+++ src/main/java/com/example/GoSonGim_BE/domain/review/dto/response/ReviewSituationDetailResponse.java
@@ -17,7 +17,7 @@ public record ReviewSituationDetailResponse(
     ) {}
     
     public record EvaluationInfo(
-        Float score,
+        Integer score,
         String feedback
     ) {}
